improve find_roots function, unit tests pass
