+++ .gitignore
@@ -1,3 +1,5 @@
 __pycache__
 example.py
 .venv
+# Ignore dynaconf secret files
+.secrets.*
