+++ settings.toml
@@ -1,0 +1,3 @@
+[polynomial]
+MAX_ROOT_INITIALIZATIONS = 10
+MAX_ATTEMPTS_PER_ROOT = 500
